m136: Update core/ docs/ modules/ and move svg dom implementation where it belongs
