+++ skia-bindings/src/bindings.cpp
@@ -1916,7 +1916,7 @@ extern "C" void C_SkContourMeasure_end(const SkContourMeasure* self, SkContourMe
 }
 
 extern "C" bool C_SkContourMeasure_ForwardVerbIterator_Equals(const SkContourMeasure::ForwardVerbIterator* a, const SkContourMeasure::ForwardVerbIterator* b) {
-    return *const_cast<SkContourMeasure::ForwardVerbIterator*>(a) == *b;
+    return *a == *b;
 }
 
 extern "C" void C_SkContourMeasure_ForwardVerbIterator_item(const SkContourMeasure::ForwardVerbIterator* self, SkContourMeasure::VerbMeasure* uninitialized) {
