+++ skia-safe/src/core/milestone.rs
@@ -1,1 +1,1 @@
-pub const MILESTONE: usize = 135;
+pub const MILESTONE: usize = 136;
